+++ Reprise.IntegrationTests/WeatherTests.cs
@@ -3,6 +3,10 @@ namespace Reprise.IntegrationTests
     public class WeatherTests : TestBase
     {
         [Fact]
-        public async Task Get() => await Verify(await Client.GetAsync("/weather")).IgnoreMembersWithType<HttpContent>();
+        public async Task Get()
+        {
+            await Verify(await Client.GetAsync("/weather"))
+                .IgnoreMembersWithType<HttpContent>();
+        }
     }
 }