+++ Reprise.SampleApi/Endpoints/GetWeather.cs
@@ -11,6 +11,7 @@ namespace Reprise.SampleApi.Endpoints
         }
 
         [Get("/weather")]
+        [AllowAnonymous]
         public static async Task<WeatherForecast[]> Handle(IWeatherService weatherService)
         {
             return await weatherService.GetForecast();