+++ Reprise.SampleApi/Usings.cs
@@ -1,1 +1,2 @@
+global using Microsoft.AspNetCore.Authorization;
 global using Reprise.SampleApi.Data;