+++ Reprise/WebApplicationBuilderExtensions.cs
@@ -8,11 +8,11 @@
         internal static TypeProcessorFactory _TypeProcessorFactory = new();
 
         /// <summary>
-        /// Sets up the DI container.
-        /// </summary>
-        /// <remarks>
-        /// This method performs an assembly scan to:
+        /// Sets up the DI container by performing an assembly scan to:
         /// <list type="bullet">
+        ///     <item>
+        ///     Discover all API endpoints (types decorated with <see cref="EndpointAttribute"/>).
+        ///     </item>
         ///     <item>
         ///     Call all <see cref="IServiceConfigurator.ConfigureServices(WebApplicationBuilder)"/> 
         ///     implementations.
@@ -22,11 +22,13 @@
         ///     and add them with a <see cref="ServiceLifetime.Singleton"/>.
         ///     </item>
         /// </list>
-        /// </remarks>
+        /// </summary>
         /// <exception cref="ArgumentNullException"/>
         /// <exception cref="InvalidOperationException"/>
-        public static WebApplicationBuilder ConfigureServices(this WebApplicationBuilder builder) =>
-            builder.ConfigureServices(Assembly.GetCallingAssembly());
+        public static WebApplicationBuilder ConfigureServices(this WebApplicationBuilder builder)
+        {
+            return builder.ConfigureServices(Assembly.GetCallingAssembly());
+        }
 
         /// <inheritdoc cref="ConfigureServices(WebApplicationBuilder)"/>
         public static WebApplicationBuilder ConfigureServices(this WebApplicationBuilder builder, Assembly assembly)
@@ -37,9 +39,9 @@
             var types = assembly.GetExportedTypes().Where(t => t.IsClass && !t.IsAbstract);
             foreach (var type in types)
             {
-                processors.ForEach(processor => processor.Process(builder, type));
+                processors.ForEach(p => p.Process(builder, type));
             }
-            processors.ForEach(processor => processor.PostProcess(builder));
+            processors.ForEach(p => p.PostProcess(builder));
 
             return builder;
         }
@@ -47,10 +49,12 @@
 
     internal class TypeProcessorFactory
     {
-        internal virtual List<AbstractTypeProcessor> CreateAll() =>
-            typeof(WebApplicationBuilderExtensions).Assembly.GetTypes()
+        internal virtual List<AbstractTypeProcessor> CreateAll()
+        {
+            return typeof(WebApplicationBuilderExtensions).Assembly.GetTypes()
                 .Where(t => t.IsClass && !t.IsAbstract && t.IsAssignableTo(typeof(AbstractTypeProcessor)))
                 .Select(t => (AbstractTypeProcessor)Activator.CreateInstance(t)!)
                 .ToList();
+        }
     }
 }